requirements.txt
PyYAML>=6.0
rich>=13.0
mcp>=0.1.0
